Remove node, and give Component more responsibility
