+++ src/ui/ui.h
@@ -1,13 +1,20 @@
 #pragma once
 
-#include "ui_node.h"
+#include "component.h"
+
 #include "../input/input.h"
 
 namespace ui {
+
+    struct UIMouseClickEvent {
+        uint32_t button;
+        glm::ivec2 mouse_pos;
+    };
+
     class UI {
     public:
         UI();
-        std::unique_ptr<ui::UINode> root;
+        std::unique_ptr<ui::UIComponent> root;
 
         void handle_mouse_button_event(const input::KeyCombination &combi);
         void handle_key_event(const input::KeyCombination &combi);
